Push async find next/prev down into LogStack
Searching next and prev could take a long time at first.  If it takes more
than 10ms, leave the operation "pending" and track the state inside
LogStack itself.  Also move the FillGaps position down into LogStack so
we no longer have to search from the start each time.

Eventually we need to get the status line details from LogStack, too,
including "Searching..." and "Search not found".

This adds a small weird anomaly where we can continue to navigate the log
until the search is found, at which point we will jump to the found item.  This
can be unexpected and weird.  Should we fix it?  Do we need a "cancel" option?
Or is this okay?

+++ indexed_file/src/indexer/indexed_log.rs
@@ -88,8 +88,8 @@ pub trait IndexedLog {
     where Self: Sized ;
 
     // Autowrap
-    fn with_timeout(&mut self, ms: usize) -> TimeoutWrapper<Self> where Self: std::marker::Sized {
-        TimeoutWrapper::new(self, ms)
+    fn with_timeout(&mut self, ms: u64) -> TimeoutWrapper<Self> where Self: std::marker::Sized {
+        TimeoutWrapper::new(self, ms as usize)
     }
 
     // Iterators
